Added Carousel
Messed around with Bootstrap built in carousel for the weekly picks.
Also changed 'Monthly Win Rate' to 'Monthly Rate'.

Carousel currently shows three images at a time, able to go left and
right. Slight transition glitch needs to be worked out. Added in a
header above the carousel, messed around with border by making a div for
the border, and putting the heading inside to get get the correct
spacing. Otherwise the border would take the full width of the page.
Added fonts folder for glyphicons to work.